+++ tests/testthat/helper-content-replay.R
@@ -1,5 +1,5 @@
-test_record_replay <- function(x, tools = list()) {
+test_record_replay <- function(x, tools = list(), .envir = parent.frame()) {
   recorded <- contents_record(x)
-  replayed <- contents_replay(recorded, tools = tools)
+  replayed <- contents_replay(recorded, tools = tools, .envir = .envir)
   expect_equal(replayed, x)
 }